+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.47.0
+
+- Added `onPatches` callback option to the `patchRecorder`.
+- The `events` list in the `patchRecorder` is now an observable array.
 - `types.enum` no longer requires the generic type in typescript, the enum object as parameter is enough.
 
 ## 0.46.0